+++ public/translations/de.json
@@ -56,8 +56,10 @@
     "home": {
         "forumIntroPart1": "Entdecke mehr Erkenntnisse, tausch dich mit Gleichgesinnten aus und bleib über die neuesten Diskussionen in unserem Forum informiert: ",
         "forumIntroPart2": ". Vertief dich in Themen, teile deine Erfahrungen und vernetze dich mit unserer Community. Tritt noch heute in die Diskussion ein!",
-        "helloTonies": {
-            "navigationTitle": "Tonieparty"
+        "404": {
+            "navigationTitle": "404 - Seite nicht gefunden",
+            "title": "404",
+            "description": "Hups - Seite nicht gefunden"
         },
         "helpfulLinks": "Hilfreiche Links",
         "intro": "TeddyCloud ist eine alternative Serverlösung für deine Toniebox, die es dir ermöglicht, Cloud-Dienste lokal zu hosten. Dadurch behältst du die Kontrolle darüber, welche Daten an die Cloud des Originalherstellers gesendet werden, und kannst deine eigenen Hörspiel-Dateien für die Figuren selbst hosten.",
