+++ public/translations/en.json
@@ -56,8 +56,10 @@
     "home": {
         "forumIntroPart1": "Discover more insights, engage with like-minded individuals, and stay updated on the latest discussions at our forum: ",
         "forumIntroPart2": ". Dive deeper into topics, share your experiences, and connect with our community. Join the conversation today!",
-        "helloTonies": {
-            "navigationTitle": "Tonieparty"
+        "404": {
+            "navigationTitle": "404 - Page not found",
+            "title": "404",
+            "description": "Oops - Page not found"
         },
         "helpfulLinks": "Helpful Links",
         "intro": "TeddyCloud is an alternative server for your Toniebox, allowing you to host the cloud services locally. This gives you the control about which data is sent to the original manufacturer’s cloud and allows you to host your own figurine audio files.",
