+++ src/App.tsx
@@ -8,7 +8,7 @@ import { SettingsPage } from "./pages/settings/SettingsPage";
 import { RtnlPage } from "./pages/settings/RtnlPage";
 import { CertificatesPage } from "./pages/settings/certificates/CertificatesPage";
 import { HomePage } from "./pages/home/HomePage";
-import { HelloToniesPage } from "./pages/home/HelloToniesPage";
+import { Error404Page } from "./pages/Error404Page";
 import { StatsPage } from "./pages/home/StatsPage";
 import { StyledHeader } from "./components/header/StyledHeader";
 import { StyledFooter } from "./components/footer/StyledFooter";
@@ -85,7 +85,6 @@ function App() {
                             <Layout>
                                 <Routes>
                                     <Route path="/" element={<HomePage />} />
-                                    <Route path="/home/hellotonies" element={<HelloToniesPage />} />
                                     <Route path="/home/stats" element={<StatsPage />} />
                                     <Route path="/tonies" element={<ToniesPage />} />
                                     <Route path="/tonies/system-sounds" element={<SystemSoundsPage />} />
@@ -107,6 +106,7 @@ function App() {
                                     <Route path="/community/contributors" element={<ContributorsPage />} />
                                     <Route path="/community/changelog" element={<ChangelogPage />} />
                                     <Route path="/uitest" element={<UiTest />} />
+                                    <Route path="/*" element={<Error404Page />} />
                                 </Routes>
                             </Layout>
                             <StyledFooter />
